requirements.txt
bidict==0.23.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-Hashing==1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
python-engineio==4.10.1
Werkzeug==3.0.3
wsproto==1.2.0
